Finally finished impelmenting imgui backend , have multiple viewports. All (apparent)  bugs fixed.

+++ UntitledGameEngine/src/UGE/core/TimeStep.h
@@ -6,18 +6,18 @@ namespace UGE {
 	class TimeStep {
 
 	public:
-		TimeStep(double time = 0.0f)
+		TimeStep(float time = 0.0f)
 			:m_Time(time)
 		{};
 
 
-		inline double getSecond() const { return m_Time; };
-		inline double getMillisecond() const { return m_Time * 1000; };
+		inline float getSecond() const { return m_Time; };
+		inline float getMillisecond() const { return m_Time * 1000; };
 
-		operator double() const { return m_Time; };
+		operator float() const { return m_Time; };
 
 	private:
-		double m_Time;
+		float m_Time;
 
 
 	};
